+++ packages/codemirror-block-editor/src/extension/add-block-on-new-line.ts
@@ -1,4 +1,5 @@
 import { ChangeSpec, EditorState } from "@codemirror/state";
+import { blockMarkerFacet } from "./block-marker-facet";
 import {
   findBlockLevelCharacterIndentationOfLine,
   findBlockLevelOfLineNumberInDocument,
@@ -8,6 +9,7 @@ export const addBlockOnNewLine = EditorState.transactionFilter.of(
   (transaction) => {
     const doc = transaction.newDoc;
     const changes: ChangeSpec[] = [];
+    const blockMarker = transaction.startState.facet(blockMarkerFacet);
     transaction.changes.iterChanges((fromA, toA, fromB, toB, text) => {
       const fromLine = doc.lineAt(fromA);
       const toLine = doc.lineAt(toB);
@@ -19,21 +21,23 @@ export const addBlockOnNewLine = EditorState.transactionFilter.of(
       ) {
         const line = doc.line(lineNumber);
         const blockLevelOfLine = findBlockLevelCharacterIndentationOfLine(
-          line.text
+          line.text,
+          blockMarker
         );
         if (blockLevelOfLine <= 0) {
           const shouldBlockLevel = findBlockLevelOfLineNumberInDocument(
             doc,
-            lineNumber
+            lineNumber,
+            blockMarker
           );
           const numOfIndentationSpaces =
             line.text.length - line.text.trimLeft().length;
-          // TODO replace magic number 2 with Block Marker length
-          const missingSpaces = shouldBlockLevel - numOfIndentationSpaces - 2;
-          if (missingSpaces > -2) {
+          const missingSpaces =
+            shouldBlockLevel - numOfIndentationSpaces - blockMarker.length;
+          if (missingSpaces > -blockMarker.length) {
             changes.push({
               from: line.from,
-              insert: " ".repeat(missingSpaces) + "- ",
+              insert: " ".repeat(missingSpaces) + blockMarker,
             });
           }
         }