+++ packages/codemirror-block-editor/CHANGELOG.md
@@ -5,6 +5,14 @@
 - Disable `Tab+ESC` escape hatch for vim compatibility
 - Improve testing pipeline by properly integrating the Karma runner
 
+## 0.8.0
+
+- Setting the block marker in facet
+
+## 0.7.7
+
+- Fixing git tagging
+
 ## 0.7.6
 
 - Manual git tagging